+++ pkg/controllers/nodeclaim/consistency/nodeshape.go
