+++ pkg/controllers/nodeclaim/consistency/termination.go
