+++ pkg/controllers/nodeclaim/lifecycle/initialization.go
