+++ pkg/controllers/nodeclaim/lifecycle/machine_registration_test.go
