+++ pkg/controllers/nodeclaim/lifecycle/nodeclaim_registration_test.go
