+++ pkg/controllers/nodeclaim/lifecycle/registration.go
