+++ pkg/controllers/nodepool/counter/controller.go
