+++ pkg/controllers/nodepool/hash/controller.go
